+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.3.1 (2013-07-22)
 ------------------
 
+- add formlib uber selection override for portlets
+  [vangheem]
+
 - handle plone.app.relationfield not being installed
   [vangheem]